+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.2.0-dev
 - only first line of sections TEST, ARGS and FILE_EXTERNAL is now parsed
 - sections GET, COOKIE and PHPDBG are parsed now
 - implemented sections EXPECTREGEX, EXPECTREGEX_EXTERNAL, CAPTURE_STDIO, EXPECTF and EXPECTF_EXTERNAL
+- tests are marked as flaky/supposed to fail if SKIPIF section has output starting with "flaky" or "xfail" respectively
 
 Version 0.1.0
 - initial version